requirements.txt
pygame_ce
pygame_gui
pytest
pytest-cov
pytest-timeout
